+++ typings/response.d.ts
@@ -143,6 +143,33 @@ export interface FetchResult {
    }[];
 }
 
+/**
+ * The `InitResult` is returned when (re)initialising a git repo.
+ */
+export interface InitResult {
+   /**
+    * Boolean representing whether the `--bare` option was used
+    */
+   readonly bare: boolean;
+
+   /**
+    * Boolean representing whether the repo already existed (re-initialised rather than initialised)
+    */
+   readonly existing: boolean;
+
+   /**
+    * The path used when initialising
+    */
+   readonly path: string;
+
+   /**
+    * The git configuration directory - for a bare repo this is the same as `path`, in non-bare repos
+    * this will usually be a sub-directory with the name `.git` (or value of the `$GIT_DIR` environment
+    * variable).
+    */
+   readonly gitDir: string;
+}
+
 export interface MoveSummary {
    moves: any[];
 }
@@ -170,6 +197,9 @@ export interface PullResult {
    deleted: string[];
 }
 
+/**
+ * Represents file name changes in a StatusResult
+ */
 export interface StatusResultRenamed {
    from: string;
    to: string;
@@ -193,6 +223,10 @@ export interface FileStatusResult {
    working_dir: string;
 }
 
+/**
+ * The StatusResult is returned for calls to `git.status()`, represents the state of the
+ * working directory.
+ */
 export interface StatusResult {
    not_added: string[];
    conflicted: string[];